feat: Updated tests
