--- packages/topic/__tests__/android/author-profile.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared";
-
-describe("AuthorProfile tests on android", shared);
