+++ packages/topic/__tests__/android/topic-functional.test.js
@@ -1,0 +1,5 @@
+import sharedStyling from "../topic-functional";
+
+describe("Topic functional tests on android", () => {
+  sharedStyling();
+});
