+++ packages/topic/__tests__/android/topic-styling.test.js
@@ -1,0 +1,5 @@
+import sharedStyling from "../topic-styling";
+
+describe("Topic styling tests on android", () => {
+  sharedStyling();
+});
