+++ packages/topic/__tests__/ios/topic-functional.test.js
@@ -1,0 +1,5 @@
+import sharedStyling from "../topic-functional";
+
+describe("Topic functional tests on ios", () => {
+  sharedStyling();
+});
