+++ packages/topic/__tests__/ios/topic-head-functional.test.js
@@ -1,5 +1,5 @@
 import { mockReactNativeComponent } from "@times-components/jest-configurator";
-import sharedFunctional from "../shared-functional";
+import sharedFunctional from "../topic-head-functional";
 
 describe("TopicHead functional tests on ios", () => {
   jest.mock("Text", () => mockReactNativeComponent("Text"));
