+++ packages/topic/__tests__/ios/topic-head-styling.test.js
@@ -1,4 +1,4 @@
-import sharedStyling from "../shared-styling";
+import sharedStyling from "../topic-head-styling";
 
 describe("TopicHead styling tests on ios", () => {
   sharedStyling();
