+++ packages/topic/__tests__/topic-head-styling.js
@@ -6,7 +6,7 @@ import TopicHead from "../src/topic-head";
 
 export default () => {
   it("should render styling correctly", () => {
-    const tree = renderer.create(<TopicHead name="Animals" />).toJSON();
+    const tree = renderer.create(<TopicHead name="Animals" description="Animals are multicellular eukaryotic organisms." />).toJSON();
 
     expect(tree).toMatchSnapshot();
   });
