+++ packages/topic/__tests__/topic-styling.js
@@ -1,0 +1,37 @@
+import "react-native";
+import "jest-styled-components";
+import React from "react";
+import renderer from "react-test-renderer";
+import { fixtureGenerator } from "@times-components/provider-test-tools";
+import { MockedProvider } from "@times-components/utils";
+import Topic from "../src/topic";
+
+export default () => {
+  const topicProps = {
+    analyticsStream: () => {},
+    onArticlePress: () => {},
+    refetch: () => {},
+    slug: "chelsea"
+  };
+
+  it("should render styling correctly", () => {
+    const pageSize = 3;
+    const tree = renderer.create(
+      <MockedProvider
+        mocks={fixtureGenerator.makeTopicArticleMocks({
+          pageSize,
+          withImages: true
+        })}
+      >
+        <Topic
+          {...topicProps}
+          analyticsStream={() => {}}
+          page={1}
+          pageSize={pageSize}
+        />
+      </MockedProvider>
+    );
+
+    expect(tree).toMatchSnapshot();
+  });
+};
