--- packages/topic/__tests__/web/author-profile.web.test.js
@@ -1,3 +1,0 @@
-import shared from "../shared";
-
-describe("AuthorProfile tests on web", shared);
