+++ packages/topic/__tests__/web/topic-head-functional.test.js
@@ -1,4 +1,4 @@
-import sharedFunctional from "../shared-functional";
+import sharedFunctional from "../topic-head-functional";
 
 describe("TopicHead functional tests on web", () => {
   sharedFunctional();
