+++ packages/topic/__tests__/web/topic-styling.test.js
@@ -1,0 +1,5 @@
+import sharedStyling from "../topic-styling";
+
+describe("Topic styling tests on web", () => {
+  sharedStyling();
+});
